Use proper test package; move common code to internal
