+++ index/percolator.go
@@ -16,6 +16,7 @@ import (
 	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgxpool"
 	"github.com/opensearch-project/opensearch-go/v2"
+	"github.com/ttab/elephant-index/internal"
 	"github.com/ttab/elephant-index/postgres"
 	"github.com/ttab/elephantine"
 	"github.com/ttab/elephantine/pg"
@@ -681,8 +682,8 @@ func (p *Percolator) percolateDocument(
 	index string,
 	doc postgres.PercolatorDocument,
 ) (outErr error) {
-	payload, err := json.Marshal(searchRequestV1{
-		Query: qWrap("percolate", percolateQuery{
+	payload, err := json.Marshal(internal.SearchRequestV1{
+		Query: internal.QWrap("percolate", percolateQuery{
 			Field:    "query",
 			Document: doc.Fields,
 		}),
